[5541] * [new sql updates 5541_mangos_command.sql 5541_mangos_mangos_string.sql] Implement .learn all_default command (for selected player) that allow re-learn lost default (added at creating and rewarded by completed quests) spells back to player without reset other spells. * At learning talent spell always unlearn other talent ranks, at learning non-talent spell always learn lower ranks before spell learn other talent ranks, at unlearn non-talents spell do unlearn higher spell ranks. This all will preserve spellbook integrity at loading and use .(un)learn commands.

+++ src/game/SpellEffects.cpp
@@ -2598,10 +2598,7 @@ void Spell::EffectLearnSpell(uint32 i)
     Player *player = (Player*)unitTarget;
 
     uint32 spellToLearn = m_spellInfo->EffectTriggerSpell[i];
-    //data.Initialize(SMSG_LEARNED_SPELL);
-    //data << spellToLearn;
-    //player->GetSession()->SendPacket(&data);
-    player->learnSpell((uint16)spellToLearn);
+    player->learnSpell(spellToLearn);
 
     sLog.outDebug( "Spell: Player %u have learned spell %u from NpcGUID=%u", player->GetGUIDLow(), spellToLearn, m_caster->GetGUIDLow() );
 }
